letter dates: render full month name ("06 June 2026" not "06 Jun 2026")
User wanted the {{DocumentFilterInfo.ShortDate}} placeholder in
the Revised Offer Letter (and every other template that uses it)
to show the full month name. Standard HR letter convention.

Flipped fmtShortDate's `month: "short"` → `month: "long"` in both
the explicit-date and current-date branches. The function name is
now a tiny lie ("short" date → long month) but renaming would
ripple through every template seed. Comment explains the legacy
naming.

+++ src/lib/hr/letter-render.ts
@@ -201,11 +201,15 @@ const fmtDate = (d: Date | null | undefined): string => {
   if (Number.isNaN(date.getTime())) return "—";
   return date.toLocaleDateString("en-IN", { day: "2-digit", month: "long", year: "numeric" });
 };
+// Despite the legacy name, fmtShortDate uses the FULL month name
+// ("06 June 2026") to match HR's preferred letter-style format.
+// Was 3-letter month previously ("06 Jun 2026"); flipped to long
+// after HR feedback on the Revised Offer Letter preview.
 const fmtShortDate = (d: Date | null | undefined): string => {
-  if (!d) return new Date().toLocaleDateString("en-IN", { day: "2-digit", month: "short", year: "numeric" });
+  if (!d) return new Date().toLocaleDateString("en-IN", { day: "2-digit", month: "long", year: "numeric" });
   const date = d instanceof Date ? d : new Date(d as any);
   if (Number.isNaN(date.getTime())) return "—";
-  return date.toLocaleDateString("en-IN", { day: "2-digit", month: "short", year: "numeric" });
+  return date.toLocaleDateString("en-IN", { day: "2-digit", month: "long", year: "numeric" });
 };
 
 // Resolves a placeholder against the employee row + custom inputs.
